backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
groq==0.3.0
numpy==1.26.4
langchain==0.1.16
playwright==1.42.0
beautifulsoup4==4.12.3
requests==2.31.0
faiss-cpu==1.7.4
torch==2.2.1
